+++ R/Task.R
@@ -145,11 +145,36 @@ Task = R6Class("Task",
       )
 
       cn = self$col_info$id # note: this sorts the columns!
-      private$.row_roles = list(use = rn, test = integer(), holdout = integer())
+      private$.row_roles = list(use = rn)
       private$.col_roles = named_list(mlr_reflections$task_col_roles[[task_type]], character())
       private$.col_roles$feature = setdiff(cn, self$backend$primary_key)
       self$extra_args = assert_list(extra_args, names = "unique")
       self$mlr3_version = mlr_reflections$package_version
+    },
+
+    #' @description
+    #' Splits a test or holdout task from the task.
+    #' This modifies the task in-place.
+    #' Subsequent operations on the (primary) task are not relayed to the test task.
+    #' @param row_ids (`integer()` or `NULL`)\cr
+    #'   The row ids to use for the test task.
+    #' @param type (`character(1)`)\cr
+    #'   The task to split off. Either `"test"` (default) or `"holdout"`.
+    #' @param remove (`logical(1)`)\cr
+    #'   If `TRUE` (default), the `row_ids` are removed from the primary task's active `"use"` rows.
+    #' @return Modified `self`.
+    partition = function(row_ids = NULL, type = "test", remove = TRUE) {
+      assert_row_ids(row_ids, null.ok = FALSE)
+      assert_subset(type, c("test", "holdout"))
+      task_name = paste0(type, "_task")
+      if (!is.null(self[[task_name]])) lg$debug("Overwriting existing %s task for task '%s'", type,  self$id)
+      on.
+      $test_task = new_task$holdout_task = NULL
+      new_task = self$clone(deep = TRUE)
+      new_task$row_roles$use = row_ids
+      self[[task_name]] = new_task
+      if (remove) self$row_roles$use = setdiff(self$row_roles$use, row_ids)
+      invisible(self)
     },
 
     #' @description
@@ -201,13 +226,11 @@ Task = R6Class("Task",
         catn(str_indent(sprintf("* %s:", str), roles[[role]]))
       })
 
-      nrows = list(test = length(private$.row_roles$test), holdout = length(private$.row_roles$holdout))
-      if (nrows$test || nrows$holdout) {
-        str = paste(c(
-          if(nrows$test) sprintf("%i (test)", nrows$test),
-          if(nrows$holdout) sprintf("%i (holdout)", nrows$holdout)
-          ), collapse = ", ")
-        catf(str_indent("* Additional Row Roles:", str))
+      if (!is.null(private$.test_task)) {
+        catf(str_indent("* Test Task:", sprintf("(%ix%i)", private$.test_task$nrow, private$.test_task$ncol)))
+      }
+      if (!is.null(private$.holdout_task)) {
+        catf(str_indent("* Holdout Task:", sprintf("(%ix%i)", private$.holdout_task$nrow, private$.holdout_task$ncol)))
       }
     },
 
@@ -342,6 +365,7 @@ Task = R6Class("Task",
     #' @return Named `integer()`.
     missings = function(cols = NULL) {
       assert_has_backend(self)
+
 
       if (is.null(cols)) {
         cols = unlist(private$.col_roles[c("target", "feature")], use.names = FALSE)
@@ -497,20 +521,16 @@ Task = R6Class("Task",
     #' Adds additional columns to the [DataBackend] stored in `$backend`.
     #'
     #' The row ids must be provided as column in `data` (with column name matching the primary key name of the [DataBackend]).
-    #' If this column is missing, it is assumed that the rows are exactly in the order of
-    #' `c(task$row_roles$use, task$row_roles$test)`.
+    #' If this column is missing, it is assumed that the rows are exactly in the order of `$row_ids`.
     #' In case of name clashes of column names in `data` and [DataBackend], columns in `data` have higher precedence
     #' and virtually overwrite the columns in the [DataBackend].
     #'
     #' This operation mutates the task in-place.
     #' See the section on task mutators for more information.
     #' @param data (`data.frame()`).
-    #' @param sets (`character()`)\cr
-    #'   For which sets to bind data. Can be a ubset of `"use"`, `"test"` and `"holdout"`.
-    cbind = function(data, sets = "use") {
+    cbind = function(data) {
       assert_has_backend(self)
       pk = self$backend$primary_key
-      assert_subset(sets, c("use", "test", "holdout"))
 
       if (is.data.frame(data)) {
         # binding data with 0 rows is explicitly allowed
@@ -518,14 +538,14 @@ Task = R6Class("Task",
           return(invisible(self))
         }
 
-        row_ids = if (pk %in% names(data)) pk else invoke(c, .args = self$row_roles[sets])
+        row_ids = if (pk %in% names(data)) pk else self$row_ids
         data = as_data_backend(data, primary_key = row_ids)
       } else {
         assert_backend(data)
         if (data$ncol <= 1L) {
           return(invisible(self))
         }
-        assert_set_equal(invoke(c, .args = self$row_roles[sets]), data$rownames)
+        assert_set_equal(self$row_ids, data$rownames)
       }
 
       # update col_info for existing columns
@@ -736,14 +756,58 @@ Task = R6Class("Task",
       private$.id = assert_string(rhs, min.chars = 1L)
     },
 
+    #' @field test_task (`Task` or `NULL`)\cr
+    #' Optional test task that can e.g. be used for early stopping with learners such as XGBoost.
+    #' When training a learner that has the property 'uses_test_task', as well as rows with the row role 'test',
+    #' this task is automatically generated.
+    test_task = function(rhs) {
+      if (missing(rhs)) {
+        return(invisible(private$.test_task))
+      }
+      if (is.null(rhs)) {
+        private$.test_task = NULL
+        return(invisible(private$.test_task))
+      }
+
+      assert_task(rhs, task_type = self$task_type)
+      if (identical(rhs, self)) { # avoid circles
+        stopf("Task '%s' cannot be its own test task", self$id)
+      }
+      if (!is.null(rhs$test_task) || !is.null(rhs$holdout_task)) { # avoid recursive structures
+        stopf("Trying to assign task '%s' as a test task, remove its test/holdout task first", rhs$id)
+      }
+      private$.test_task = rhs
+      invisible(private$.test_task)
+    },
+
+    #' @field test_task (`Task` or `NULL`)\cr
+    #' Optional holdout task.
+    #' It is possible to make predictions on this task, by setting the `predict_set`
+    #' When training a learner that has the property 'uses_test_task', as well as rows with the row role 'test',
+    #' this task is automatically generated.
+    holdout_task = function(rhs) {
+      if (missing(rhs)) {
+        return(invisible(private$.holdout_task))
+      }
+      if (missing(rhs)) {
+        private$.holdout_task = NULL
+        return(invisible(private$.holdout_task))
+      }
+      assert_task(rhs, task_type = self$task_type)
+      if (identical(rhs, self)) {
+        stopf("Task '%s' cannot be its own holdout task", self$id)
+      }
+      if (!is.null(rhs$test_task) || !is.null(rhs$holdout_task)) { # avoid recursive structures
+        stopf("Trying to assign task '%s' as a holdout task, remove its test/holdout task first", rhs$id)
+      }
+      private$.holdout_task = rhs
+      invisible(private$.holdout_task)
+    },
 
     #' @template field_hash
     hash = function(rhs) {
       if (is.null(private$.hash)) {
-        private$.hash = calculate_hash(
-          class(self), self$id, self$backend$hash, self$col_info,
-          remove_named(private$.row_roles, "test"), private$.col_roles, private$.properties
-        )
+        private$.hash = task_hash(self, self$row_roles)
       }
 
       private$.hash
@@ -836,9 +900,11 @@ Task = R6Class("Task",
 
       assert_has_backend(self)
       assert_list(rhs, .var.name = "row_roles")
+      if ("test" %in% names(rhs) || "holdout" %in% names(rhs)) {
+        stopf("Setting row roles 'test'/'holdout' is no longer possible, use $split_<test/holdout>() instead")
+      }
       assert_names(names(rhs), "unique", permutation.of = mlr_reflections$task_row_roles, .var.name = "names of row_roles")
       rhs = map(rhs, assert_row_ids, .var.name = "elements of row_roles")
-
       private$.hash = NULL
       private$.row_roles = rhs
     },
@@ -1028,6 +1094,8 @@ Task = R6Class("Task",
   ),
 
   private = list(
+    .holdout_task = NULL,
+    .test_task = NULL,
     .id = NULL,
     .properties = NULL,
     .col_roles = NULL,
@@ -1037,7 +1105,13 @@ Task = R6Class("Task",
 
     deep_clone = function(name, value) {
       # NB: DataBackends are never copied!
-      if (name == "col_info") copy(value) else value
+      if (name == "col_info") {
+        copy(value)
+      } else if (name %in% c(".test_task", ".holdout_task") && !is.null(value)) {
+        value$clone(deep = TRUE)
+      } else {
+        value
+      }
     }
   )
 )
@@ -1158,6 +1232,8 @@ task_rm_backend = function(task) {
   ee = get_private(task)
   ee$.hash = force(task$hash)
   ee$.col_hashes = force(task$col_hashes)
+  ee$.test_task$backend = NULL
+  ee$.holdout_task$backend = NULL
 
   # NULL backend
   task$backend = NULL